feat: init PaymentSelectorBottom
