+++ src/components/bottom-sheet/index.ts
@@ -1,1 +1,4 @@
+import BottomSheet from './BottomSheet';
+
+export { BottomSheet };
 export * from './useBottomSheet';
